Refactor/Validation-check
