+++ src/components/post/write/PostWriteRecipe.tsx
@@ -1,4 +1,4 @@
-import React from 'react';
+import React, { useEffect } from 'react';
 import { useState, useRef } from 'react';
 import { useNavigate } from 'react-router';
 import { useAtom } from 'jotai';
@@ -26,50 +26,69 @@ const PostWriteRecipe = ({ orgPostId, setCategory }: OrgPostIdProps) => {
   const [selectedImages, setImagesDataAtom] = useAtom(imagesAtom);
 
   const [title, setTitle] = useState<string>('');
+  const [isIn, setIsIn] = useState(true);
+
+  useEffect(() => {
+    const handleBeforeUnload = (e: BeforeUnloadEvent) => {
+      if (isIn) {
+        e.preventDefault();
+        e.returnValue = '작성 중인 내용이 사라집니다. 페이지를 떠나시겠습니까?';
+      }
+    };
+
+    window.addEventListener('beforeunload', handleBeforeUnload);
+
+    return () => {
+      // 컴포넌트가 언마운트될 때 이벤트 리스너 제거
+      window.removeEventListener('beforeunload', handleBeforeUnload);
+    };
+  }, [isIn]);
 
   const submitPost = async (e: React.FormEvent<HTMLFormElement>) => {
     e.preventDefault();
-    // const Message = window.confirm('글을 작성하시겠습니까?');
+    const Message = window.confirm('글을 작성하시겠습니까?');
 
-    // if (Message) {
-    if (Object.values(allContents).length === 0) {
-      alert('내용을 입력해 주세요!');
-      return;
-    }
-
-    const imageUrls = [];
-
-    for (const selectedImage of Object.values(selectedImages)) {
-      const { data, error } = await supabase.storage.from('photos').upload(`tags/${selectedImage.name}`, selectedImage);
-
-      if (error) {
-        console.error('Error uploading image to Supabase storage:', error);
-        alert('이미지 업로드 중 에러가 발생했습니다!');
+    if (Message) {
+      if (Object.values(allContents).length === 0) {
+        alert('내용을 입력해 주세요!');
         return;
       }
 
-      imageUrls.push(data.path);
+      const imageUrls = [];
+
+      for (const selectedImage of Object.values(selectedImages)) {
+        const { data, error } = await supabase.storage
+          .from('photos')
+          .upload(`tags/${selectedImage.name}`, selectedImage);
+
+        if (error) {
+          console.error('Error uploading image to Supabase storage:', error);
+          alert('이미지 업로드 중 에러가 발생했습니다!');
+          return;
+        }
+
+        imageUrls.push(data.path);
+      }
+
+      const newPost = {
+        orgPostId,
+        postCategory: 'recipe',
+        userId,
+        title,
+        body: allContents,
+        recipeBody: Object.values(allContents),
+        tags: Object.values(allTags),
+        tagimage: imageUrls
+      };
+
+      addRecipePostMutate.mutate(newPost);
+
+      setContentsAtom({});
+      setTagsDataAtom({});
+      setImagesDataAtom({});
+
+      navigate(`/`);
     }
-
-    const newPost = {
-      orgPostId,
-      postCategory: 'recipe',
-      userId,
-      title,
-      body: allContents,
-      recipeBody: Object.values(allContents),
-      tags: Object.values(allTags),
-      tagimage: imageUrls
-    };
-
-    addRecipePostMutate.mutate(newPost);
-
-    setContentsAtom({});
-    setTagsDataAtom({});
-    setImagesDataAtom({});
-
-    navigate(`/`);
-    // }
   };
 
   const clickLogo = () => {
