+++ src/components/renderPosts/reactionSource/ContentBox.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { useLocation } from 'react-router';
 
 // import { S } from '../style/StyledPostDetail';
-import TagImage from 'src/components/ImageTag/TagImage';
+import TagImage from 'src/components/ImageTag/ShowTag';
 
 import styled, { css } from 'styled-components';
